requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy==1.24.2
pandas==1.5.3
opencv-python==4.7.0.68
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
loguru==0.6.0
pytorch_lightning==1.9.2
neptune-client==0.16.18
timm==0.8.15.dev0
tensorboard==2.12.0
open_clip_torch==2.16.0
transformers==4.26.1
albumentations==1.3.0

